[DOP-4437]: Refactor project layout for plugin
